Project Added to the repository
This is the final version of the project that was submitted as a part of
requirement fulfillment for course COMS 572 - Principles of Artificial
Intelligence

This has some known issues. Like the automatic solver ( the one without
the arguments ) runs into an infiinite loop on an exception and turns
80% or 90% of the board green. Probable cause is the workarounds to
handle each and every case in a specific manner to get around a issue
(could be in design or implementation ) owing to the less amount of time
that was available. Has a lot of scope towards improvement. But serves
the purpose. This is not a new idea anyhow.